refactor: rework project structure and routing logic
