+++ src/lib/formsClient/types/formTemplate.ts
@@ -1,0 +1,3 @@
+export interface FormTemplate {
+  jsonConfig: Record<string, unknown>;
+}
