+++ src/lib/idp/introspectionCache.ts
@@ -1,6 +1,6 @@
 import { createHash } from "node:crypto";
 import type { IntrospectionResult } from "@lib/idp/introspectToken.js";
-import { RedisConnector } from "@lib/connectors/redisConnector.js";
+import { RedisConnector } from "@lib/integration/redisConnector.js";
 
 const cacheExpiry = 300; // seconds
