+++ src/lib/integration/awsServicesConnector.ts
@@ -1,8 +1,8 @@
 import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
 import { DynamoDBDocumentClient } from "@aws-sdk/lib-dynamodb";
-import { AWS_REGION, LOCALSTACK_ENDPOINT } from "@src/config.js";
-import { SecretsManagerClient } from "@aws-sdk/client-secrets-manager";
 import { SQSClient } from "@aws-sdk/client-sqs";
+import { SecretsManagerClient } from "@aws-sdk/client-secrets-manager";
+import { AWS_REGION, LOCALSTACK_ENDPOINT } from "@src/config.js";
 
 const globalConfig = {
   region: AWS_REGION,
