+++ src/lib/vault/reportProblemWithFormSubmission.ts
@@ -1,12 +1,12 @@
 import { UpdateCommand } from "@aws-sdk/lib-dynamodb";
-import { AwsServicesConnector } from "@lib/connectors/awsServicesConnector.js";
+import { AwsServicesConnector } from "@lib/integration/awsServicesConnector.js";
 import {
   FormSubmissionAlreadyReportedAsProblematicException,
   FormSubmissionNotFoundException,
-} from "@lib/vault/dataStructures/exceptions.js";
+} from "@lib/vault/types/exceptions.js";
+import { FormSubmissionStatus } from "@lib/vault/types/formSubmission.js";
 import { getFormSubmission } from "@lib/vault/getFormSubmission.js";
-import { FormSubmissionStatus } from "@lib/vault/dataStructures/formSubmission.js";
-import { logMessage } from "@lib/logger.js";
+import { logMessage } from "@lib/logging/logger.js";
 
 export async function reportProblemWithFormSubmission(
   formId: string,
