+++ src/lib/vault/types/exceptions.ts
