+++ src/middleware/globalErrorHandler/middleware.ts
@@ -1,12 +1,12 @@
 import type { Request, Response, NextFunction } from "express";
-import { logMessage } from "@src/lib/logger.js";
+import { logMessage } from "@lib/logging/logger.js";
 
 export function globalErrorHandlerMiddleware(
   error: Error,
   _request: Request,
   response: Response,
   _next: NextFunction,
-) {
+): void {
   logMessage.error(JSON.stringify(error, Object.getOwnPropertyNames(error)));
-  return response.sendStatus(500);
+  response.sendStatus(500);
 }
