--- src/routes/status/router.ts
@@ -1,9 +1,0 @@
-import { type Response, Router } from "express";
-
-export const statusApiRoute = Router();
-
-statusApiRoute.get("/", (_, response: Response) => {
-  return response.json({
-    status: "running",
-  });
-});
